entrypoint.sh: use "service restart" due to leftover PID files
When docker container is being built, starting memcached in RUN line
(which is necessary for install.php) creates a PID file.

Unfortunately "service memcached start" doesn't check that lock file
for being stale (not containing a PID of a running memcached process),
which results in Memcached not starting.

"service memcached restart" will be used instead.
To avoid similar issues, "restart" will be used for other services too.
(note that none of them are running at this point, so it's equivalent
to "start" for them)